added several orm adaptors
