+++ lib/classy_enum/adaptor/active_record.rb
@@ -1,0 +1,3 @@
+require 'classy_enum/adaptor/base'
+
+ActiveRecord::Base.send :extend, ClassyEnum::OrmAdaptor
