+++ lib/classy_enum/adaptor/mongo_mapper.rb
@@ -1,0 +1,3 @@
+require 'classy_enum/adaptor/base'
+
+MongoMapper::Document.send :include, ClassyEnum::OrmAdaptor
