+++ lib/classy_enum/adaptor/mongoid.rb
@@ -1,0 +1,3 @@
+require 'classy_enum/adaptor/base'
+
+Mongoid::Document.send :include, ClassyEnum::Adaptor::Base
